feat: C runtime externs, TLS 1.2 server, interpreter handlers
- C runtime: rt_byte_char, rt_sha1, rt_http_get, rt_random_i64, rt_base64url_*,
  rt_time_now_seconds, rt_text_to_bytes, rt_bytes_to_text (310 lines)
- TLS 1.2 server: full handshake state machine, AES-GCM record encryption,
  AsyncTlsStream/AsyncTlsListener using real building blocks (835 lines)
- Rust interpreter: 10 new handlers (random, time, byte_char, sha1, base64url,
  http_get, ws stubs) + fix rt_time_now_seconds returning Float→Int

+++ src/compiler_rust/compiler/src/interpreter_extern/time.rs
@@ -17,7 +17,7 @@ use crate::value::Value;
 pub fn rt_time_now_seconds(_args: &[Value]) -> Result<Value, CompileError> {
     unsafe {
         let time = simple_runtime::value::rt_time_now_seconds();
-        Ok(Value::Float(time))
+        Ok(Value::Int(time as i64))
     }
 }
